Move `osc` folder to `controller`
